# swagger.json
{
  "openapi": "3.0.1",
  "info": {
    "title": "DataHuBI",
    "version": "v1"
  },
  "paths": {
    "/Area": {
      "get": {
        "tags": [
          "Area"
        ],
        "responses": {
          "200": {
            "description": "Success"
          }
        }
      },
      "post": {
        "tags": [
          "Area"
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/Area"
              }
            },
            "text/json": {
              "schema": {
                "$ref": "#/components/schemas/Area"
              }
            },
            "application/*+json": {
              "schema": {
                "$ref": "#/components/schemas/Area"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Success"
          }
        }
      },
      "delete": {
        "tags": [
          "Area"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "query",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success"
          }
        }
      }
    },
    "/Calculo": {
      "get": {
        "tags": [
          "Calculo"
        ],
        "responses": {
          "200": {
            "description": "Success"
          }
        }
      },
      "post": {
        "tags": [
          "Calculo"
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/Calculo"
              }
            },
            "text/json": {
              "schema": {
                "$ref": "#/components/schemas/Calculo"
              }
            },
            "application/*+json": {
              "schema": {
                "$ref": "#/components/schemas/Calculo"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Success"
          }
        }
      },
      "delete": {
        "tags": [
          "Calculo"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "query",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success"
          }
        }
      }
    },
    "/Dados": {
      "get": {
        "tags": [
          "Dados"
        ],
        "parameters": [
          {
            "name": "area",
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "resultado",
            "in": "query",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success"
          }
        }
      }
    },
    "/Dados/upload": {
      "post": {
        "tags": [
          "Dados"
        ],
        "parameters": [
          {
            "name": "area",
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "fonte",
            "in": "query",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "multipart/form-data": {
              "schema": {
                "type": "object",
                "properties": {
                  "arquivo": {
                    "type": "string",
                    "format": "binary"
                  }
                }
              },
              "encoding": {
                "arquivo": {
                  "style": "form"
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Success"
          }
        }
      }
    },
    "/Dados/processar": {
      "post": {
        "tags": [
          "Dados"
        ],
        "parameters": [
          {
            "name": "area",
            "in": "query",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success"
          }
        }
      }
    },
    "/Importacao": {
      "get": {
        "tags": [
          "Importacao"
        ],
        "responses": {
          "200": {
            "description": "Success"
          }
        }
      },
      "post": {
        "tags": [
          "Importacao"
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/Importacao"
              }
            },
            "text/json": {
              "schema": {
                "$ref": "#/components/schemas/Importacao"
              }
            },
            "application/*+json": {
              "schema": {
                "$ref": "#/components/schemas/Importacao"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Success"
          }
        }
      },
      "delete": {
        "tags": [
          "Importacao"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "query",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success"
          }
        }
      }
    },
    "/Indicativo": {
      "get": {
        "tags": [
          "Indicativo"
        ],
        "responses": {
          "200": {
            "description": "Success"
          }
        }
      },
      "post": {
        "tags": [
          "Indicativo"
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/Indicativo"
              }
            },
            "text/json": {
              "schema": {
                "$ref": "#/components/schemas/Indicativo"
              }
            },
            "application/*+json": {
              "schema": {
                "$ref": "#/components/schemas/Indicativo"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Success"
          }
        }
      },
      "delete": {
        "tags": [
          "Indicativo"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "query",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success"
          }
        }
      }
    },
    "/Metrica": {
      "get": {
        "tags": [
          "Metrica"
        ],
        "responses": {
          "200": {
            "description": "Success"
          }
        }
      },
      "post": {
        "tags": [
          "Metrica"
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/Metrica"
              }
            },
            "text/json": {
              "schema": {
                "$ref": "#/components/schemas/Metrica"
              }
            },
            "application/*+json": {
              "schema": {
                "$ref": "#/components/schemas/Metrica"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Success"
          }
        }
      },
      "delete": {
        "tags": [
          "Metrica"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "query",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success"
          }
        }
      }
    }
  },
  "components": {
    "schemas": {
      "Area": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "nullable": true
          },
          "nome": {
            "type": "string",
            "nullable": true
          },
          "cor": {
            "type": "string",
            "nullable": true
          }
        },
        "additionalProperties": false
      },
      "Calculo": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "nullable": true
          },
          "nome": {
            "type": "string",
            "nullable": true
          },
          "script": {
            "type": "string",
            "nullable": true
          },
          "parametrosEntrada": {
            "type": "string",
            "nullable": true
          },
          "arquivoSaida": {
            "type": "string",
            "nullable": true
          }
        },
        "additionalProperties": false
      },
      "CampoIndicativo": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "nullable": true
          },
          "metricaId": {
            "type": "string",
            "nullable": true
          },
          "campo": {
            "type": "string",
            "nullable": true
          }
        },
        "additionalProperties": false
      },
      "DicionarioCampos": {
        "type": "object",
        "properties": {
          "campo": {
            "type": "string",
            "nullable": true
          },
          "campoImportado": {
            "type": "string",
            "nullable": true
          }
        },
        "additionalProperties": false
      },
      "Importacao": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "nullable": true
          },
          "nome": {
            "type": "string",
            "nullable": true
          },
          "indicativoId": {
            "type": "string",
            "nullable": true
          },
          "indicativo": {
            "$ref": "#/components/schemas/Indicativo"
          },
          "campos": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/DicionarioCampos"
            },
            "nullable": true
          }
        },
        "additionalProperties": false
      },
      "Indicativo": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "nullable": true
          },
          "nome": {
            "type": "string",
            "nullable": true
          },
          "areaId": {
            "type": "string",
            "nullable": true
          },
          "area": {
            "$ref": "#/components/schemas/Area"
          },
          "campos": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/CampoIndicativo"
            },
            "nullable": true
          }
        },
        "additionalProperties": false
      },
      "Metrica": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "nullable": true
          },
          "nome": {
            "type": "string",
            "nullable": true
          }
        },
        "additionalProperties": false
      }
    }
  }
}